+++ packages/contracts/src/abis/arbOutbox.ts
@@ -1,5 +1,5 @@
 // https://etherscan.io/address/0x0B9857ae2D4A3DBe74ffE1d7DF045bb7F96E4840
-export const OUTBOX_ABI = [
+export const ARB_OUTBOX_ABI = [
   { inputs: [], name: 'AlreadyInit', type: 'error' },
   {
     inputs: [{ internalType: 'uint256', name: 'index', type: 'uint256' }],
