+++ packages/contracts/src/abis/index.ts
@@ -4,12 +4,12 @@ export * from './arbSys'
 export * from './bep20'
 export * from './erc20'
 export * from './farmingAbi'
-export * from './foxStakingV1'
+export * from './foxStaking'
 export * from './foxyStaking'
 export * from './l1ArbitrumGateway'
 export * from './l1OrbitCustomGateway'
 export * from './l2ArbitrumGateway'
-export * from './outbox'
+export * from './arbOutbox'
 export * from './uniswapV3Quoter'
 export * from './uniswapV3PoolAbi'
 export * from './rfox'
